feat: add second part of Table component,
